requirements.txt
python-dotenv==1.0.0
openai==0.27.8
psycopg2-binary==2.9.6
torch>=2.0.1
transformers>=4.34.0
sentence-transformers>=2.2.2